% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atp_2019}
\alias{atp_2019}
\title{ATP matches in 2019}
\format{
An object of class \code{"data.frame"}.
}
\source{
Tennis archive from \href{http://www.tennis-data.co.uk/}{http://www.tennis-data.co.uk/}
}
\usage{
data(atp_2019)
}
\description{
Tennis data for male matches played in 2019. Details can be found on \href{http://www.tennis-data.co.uk/notes.txt}{http://www.tennis-data.co.uk/notes.txt}
}
\examples{
head(atp_2019)
str(atp_2019)
}
\keyword{datasets}
