% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tennis_data}
\alias{tennis_data}
\title{Download tennis data}
\usage{
tennis_data(YEAR, Circuit)
}
\arguments{
\item{YEAR}{Year to consider, in \code{"YYYY"} format.}

\item{Circuit}{Tennis circuit. Valid choices are \code{"ATP"} and \code{"WTA"}.}
}
\value{
A data frame containing the data for the selected year and circuit.
}
\description{
Imports ATP or WTA data from the website
\url{http://www.tennis-data.co.uk/}.
}
\details{
The structure of the datasets provided by
\url{http://www.tennis-data.co.uk/} may vary across years.
Therefore, before combining data from multiple seasons, users should
verify that the variables are consistent across datasets and, if
necessary, retain only the common columns.
}
\examples{
\dontrun{
db <- tennis_data("2022", "ATP")
head(db)
}

}
