% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tennis_data}
\alias{tennis_data}
\title{Download data from http://www.tennis-data.co.uk/}
\usage{
tennis_data(YEAR, Circuit)
}
\arguments{
\item{YEAR}{Year to consider, in "YYYY" format. Only years from 2013 onwards are allowed}

\item{Circuit}{Valid choices for Circuit are: "ATP" or "WTA"}
}
\value{
Data.frame for the YEAR and Circuit specified
}
\description{
Imports ATP or WTA data from the site http://www.tennis-data.co.uk/
}
\examples{
db<-tennis_data("2022","ATP") 
head(db)
}
